backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
ultralytics==8.0.196
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.0.3